Rename the ValidationResultService as appropriate
