+++ src/test/java/uk/ac/ebi/subs/validator/flipper/StatusFlipperValidationResultServiceTest.java
@@ -9,12 +9,12 @@ import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import uk.ac.ebi.subs.validator.config.MongoDBDependentTest;
 import uk.ac.ebi.subs.validator.data.AggregatorToFlipperEnvelope;
 import uk.ac.ebi.subs.validator.data.SingleValidationResult;
 import uk.ac.ebi.subs.validator.data.ValidationResult;
 import uk.ac.ebi.subs.validator.data.structures.GlobalValidationStatus;
 import uk.ac.ebi.subs.validator.data.structures.ValidationAuthor;
-import uk.ac.ebi.subs.validator.flipper.config.MongoDBDependentTest;
 import uk.ac.ebi.subs.validator.repository.ValidationResultRepository;
 
 import java.util.ArrayList;
@@ -33,14 +33,14 @@ import static org.junit.Assert.assertThat;
 @EnableMongoRepositories(basePackageClasses = ValidationResultRepository.class)
 @Category(MongoDBDependentTest.class)
 @EnableAutoConfiguration
-@SpringBootTest(classes = ValidationResultService.class)
-public class ValidationResultServiceTest {
+@SpringBootTest(classes = StatusFlipperValidationResultService.class)
+public class StatusFlipperValidationResultServiceTest {
 
     @Autowired
     ValidationResultRepository repository;
 
     @Autowired
-    ValidationResultService service;
+    StatusFlipperValidationResultService service;
 
     private ValidationResult existingValidationResult;
     private AggregatorToFlipperEnvelope envelope;
