--- src/test/java/uk/ac/ebi/subs/validator/flipper/config/MongoDBDependentTest.java
@@ -1,7 +1,0 @@
-package uk.ac.ebi.subs.validator.flipper.config;
-
-/**
- * Created by rolando on 03/08/2017.
- */
-public interface MongoDBDependentTest {
-}
